refacto(backend): new backend WhiteboardInfoBackendService
* WhiteboardInfo set private inside this module
